requirements-crewai.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiohttp-retry==2.9.1
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
async-lru==2.3.0
async-timeout==4.0.1
attrs==26.1.0
babel==2.18.0
backcall==0.2.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.10.0
beniget==0.4.1
bleach==6.3.0
blinker==1.4
Brotli==1.0.9
build==1.4.2
certifi==2026.2.25
cffi==2.0.0
chardet==4.0.0
charset-normalizer==3.4.6
chromadb==1.1.1
click==8.1.8
colorama==0.4.4
coloredlogs==15.0.1
comm==0.2.3
configobj==5.0.6
crewai==1.12.2
cryptography==46.0.6
css-parser==1.0.7
cssselect==1.1.0
cssutils==1.0.2
cycler==0.11.0
dataclasses-json==0.6.7
debugpy==1.8.20
decorator==4.4.2
defusedxml==0.7.1
deprecation==2.1.0
diskcache==5.6.3
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.1.0
docstring_parser==0.17.0
durationpy==0.10
et_xmlfile==2.0.0
exceptiongroup==1.3.1
eyeD3==0.8.10
fastjsonschema==2.21.2
feedparser==6.0.8
filelock==3.25.2
flatbuffers==25.12.19
fonttools==4.29.1
fqdn==1.5.1
frozenlist==1.8.0
fs==2.4.12
fsspec==2026.3.0
gast==0.5.2
geojson==3.2.0
googleapis-common-protos==1.73.1
grandalf==0.8
greenlet==3.3.2
groq==0.37.1
grpcio==1.80.0
h11==0.16.0
hf-xet==1.4.3
html2text==2020.1.16
html5-parser==0.4.10
html5lib==1.1
httpcore==1.0.9
httplib2==0.20.2
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.8.0
humanfriendly==10.0
idna==3.3
IMDbPY==2021.4.18
importlib_metadata==8.7.1
importlib_resources==6.5.2
instructor==1.14.5
ipykernel==7.2.0
ipython==7.31.1
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.3
json5==0.10.0
json_repair==0.25.3
jsonpatch==1.33
jsonpointer==3.1.1
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.1
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
keyring==23.5.0
kiwisolver==3.5.0
kubernetes==35.0.0
lance-namespace==0.6.1
lance-namespace-urllib3-client==0.6.1
lancedb==0.30.2
langchain==1.2.14
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.23
langchain-experimental==0.4.1
langchain-groq==1.1.2
langchain-huggingface==1.2.1
langchain-ollama==1.0.1
langchain-openai==1.1.12
langchain-pinecone==0.2.13
langchain-text-splitters==1.1.1
langgraph==1.1.4
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.23
lark==1.3.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
linkify-it-py==2.1.0
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
markdown-it-py==4.0.0
MarkupSafe==2.0.1
marshmallow==3.26.2
matplotlib==3.5.1
matplotlib-inline==0.1.3
mcp==1.26.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
mechanize==0.4.7
mistune==3.2.0
mmh3==5.2.1
mock==4.0.3
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.0.3
multidict==6.7.1
mutagen==1.45.1
mypy_extensions==1.1.0
natsort==8.0.2
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
netaddr==0.8.0
networkx==3.4.2
notebook==7.5.5
notebook_shim==0.2.4
numpy==2.2.6
oauthlib==3.2.0
ollama==0.6.1
onnxruntime==1.23.2
openai==2.30.0
openpyxl==3.1.5
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
orjson==3.11.8
ormsgpack==1.12.2
overrides==7.7.0
packaging==24.2
pandocfilters==1.5.1
parso==0.8.1
pbr==5.8.0
pdfminer.six==20251230
pdfplumber==0.11.9
pexpect==4.8.0
pickleshare==0.7.5
pillow==12.1.1
pinecone==7.3.0
pinecone-client==6.0.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
platformdirs==4.9.4
ply==3.11
portalocker==2.7.0
posthog==5.4.0
prometheus_client==0.24.1
prompt-toolkit==3.0.28
propcache==0.4.1
protobuf==5.29.6
psutil==5.9.0
ptyprocess==0.7.0
py7zr==0.11.3+dfsg
pyarrow==23.0.1
pybase64==1.4.3
pycparser==3.0
pycryptodomex==3.11.0
pydantic==2.11.10
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyelftools==0.27
Pygments==2.20.0
PyJWT==2.12.1
pymacaroons==0.13.0
PyNaCl==1.5.0
pyowm==3.5.0
pyparsing==2.4.7
pypdfium2==5.6.0
PyPika==0.51.1
pyproject_hooks==1.2.0
pyRFC3339==1.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==4.1.0
python-magic==0.4.24
python-multipart==0.0.22
pythran==0.10.0
pytz==2022.1
PyYAML==6.0.3
pyzmq==27.1.0
qrcode==7.3.1
QtPy==2.0.0
referencing==0.37.0
regex==2026.1.15
reportlab==3.6.8
repoze.lru==0.7
requests==2.33.1
requests-file==1.5.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.3.3
Routes==2.5.1
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
scour==0.38.2
SecretStorage==3.3.1
Send2Trash==2.1.0
sentence-transformers==5.3.0
setproctitle==1.2.2
sgmllib3k==1.0.0
shellingham==1.5.4
simplejson==3.17.6
simsimd==6.5.16
six==1.16.0
sniffio==1.3.1
soupsieve==2.3.1
SQLAlchemy==2.0.48
sse-starlette==3.3.4
starlette==1.0.0
sympy==1.14.0
tenacity==9.1.4
terminado==0.18.1
texttable==1.6.4
textual==8.2.1
threadpoolctl==3.6.0
tiktoken==0.12.0
tinycss2==1.1.1
tldextract==3.1.2
tokenizers==0.22.2
tomli==2.0.2
tomli_w==1.1.0
torch==2.11.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
transformers==5.4.0
triton==3.6.0
typer==0.23.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uc-micro-py==2.0.0
ufoLib2==0.13.1
unicodedata2==14.0.0
Unidecode==1.3.3
uri-template==1.3.0
urllib3==1.26.5
uuid_utils==0.14.1
uv==0.9.30
uvicorn==0.42.0
uvloop==0.22.1
voila==0.5.11
wadllib==1.3.6
watchfiles==1.1.1
wcwidth==0.2.5
webcolors==25.10.0
webencodings==0.5.1
WebOb==1.8.6
websocket-client==1.9.0
websockets==16.0
widgetsnbextension==4.0.15
xlrd==1.2.0
xxhash==3.6.0
yarl==1.23.0
yt-dlp==2022.4.8
zipp==3.23.0
zstandard==0.25.0
